backend/requirements.txt
# ============================================================
# Talk2Tables — Backend Requirements
# ============================================================
# Install: pip install -r requirements.txt
# Python 3.11+ required
# ============================================================

# ── Web Framework ────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9        # For file upload (schema docs)

# ── AI / LangGraph ───────────────────────────────────────────
langgraph==0.2.28
langchain-core==0.3.15
httpx==0.27.2                  # HTTP client used by all LLM providers

# ── SQL / Database ────────────────────────────────────────────
sqlalchemy==2.0.35
sqlparse==0.5.1                # SQL parsing + validation

# DB drivers — install only the ones you need:
pymysql==1.1.1                 # MySQL 8 + MariaDB
psycopg2-binary==2.9.9         # PostgreSQL 15
# pyodbc==5.1.0                # SQL Server (requires ODBC driver installed)
# cx-Oracle==8.3.0             # Oracle (requires Oracle Instant Client)
# aiosqlite==0.20.0            # SQLite async support

# ── Auth ──────────────────────────────────────────────────────
PyJWT==2.9.0
passlib[bcrypt]==1.7.4

# ── Data / Export ─────────────────────────────────────────────
pandas==2.2.2
openpyxl==3.1.5                # Excel export

# ── Schema Doc Extraction ─────────────────────────────────────
pdfplumber==0.11.4             # PDF text extraction
python-docx==1.1.2             # Word doc extraction

# ── Validation ────────────────────────────────────────────────
pydantic==2.9.2                # Request/response models (FastAPI uses this)
pydantic-settings==2.5.2       # Settings from .env

# ── Utilities ─────────────────────────────────────────────────
python-dotenv==1.0.1           # .env file loading
tenacity==9.0.0                # Retry logic (optional enhancement)

# ── Testing ───────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2                  # Test client for FastAPI

# ── Dev Tools ─────────────────────────────────────────────────
# ruff==0.6.9                  # Linter (optional)
# black==24.10.0               # Formatter (optional)